+++ docs/source/Changelog.rst
@@ -104,6 +104,9 @@ Changed
    -  W058 Missing KiCad main config file
    -  W058 Missing default system table
 
+-  No more reports on PCB/SCH fields mismatch when the variants modified
+   the field
+
 [1.9.0] - 2026-05-12
 --------------------